update_ebpf_socket: Rename file
