+++ includes/netdata_ebpf.h
@@ -19,8 +19,8 @@ This header has the common definitions for all `.c` files.
 #include "netdata_fs.h"
 #include "netdata_hardirq.h"
 #include "netdata_mount.h"
-#include "netdata_network.h"
 #include "netdata_process.h"
+#include "netdata_socket.h"
 #include "netdata_sync.h"
 #include "netdata_swap.h"
 #include "netdata_vfs.h"
